Made orderdetails to save and redirect to payment gateway

+++ DataAccess/Data/RoomOrderDetails.cs
@@ -22,7 +22,7 @@ namespace DataAccess.Data
         public DateTime ActualCheckInDate { get; set; }
         public DateTime ActualCheckOutDate { get; set; }
         [Required]
-        public long TotalCost { get; set; }
+        public double TotalCost { get; set; }
         [Required]
         public int RoomId { get; set; }
         [Required]
